# backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Note API
paths:
  /customers:
    post:
      summary: 顧客を登録する
      operationId: postCreateCustomer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                id:
                  type: integer
                  format: int64
                title:
                  type: string
                content:
                  type: string
              required:
                - id
                - title
                - content
        required: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  id:
                    type: integer
                    format: int64
                  title:
                    type: string
                  content:
                    type: string
                required:
                  - id
                  - title
                  - content
              headers:
                Access-Control-Allow-Origin:
                  schema:
                    type: string
                Access-Control-Allow-Methods:
                  schema:
                    type: string
                Access-Control-Allow-Headers:
                  schema:
                    type: string
  '/customers/{customerID}':
    get:
      summary: 顧客を取得する
      operationId: getCustomerByID
      parameters:
        - name: customerID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  id:
                    type: integer
                    format: int64
                  title:
                    type: string
                  content:
                    type: string
                required:
                  - id
                  - title
                  - content
              headers:
                Access-Control-Allow-Origin:
                  schema:
                    type: string
                Access-Control-Allow-Methods:
                  schema:
                    type: string
                Access-Control-Allow-Headers:
                  schema:
                    type: string